feat: signed max lookup
