+++ jolt-core/src/jolt_onnx/instruction/max.rs
@@ -1,0 +1,231 @@
+use crate::{
+    field::JoltField,
+    jolt::subtable::{
+        eq_abs::EqAbsSubtable, left_msb::LeftMSBSubtable, lt_abs::LtAbsSubtable,
+        right_msb::RightMSBSubtable,
+    },
+    subprotocols::sparse_dense_shout::PrefixSuffixDecomposition,
+    utils::{instruction_utils::concatenate_lookups, uninterleave_bits},
+};
+use ark_std::log2;
+use rand::prelude::StdRng;
+use rand::RngCore;
+use serde::{Deserialize, Serialize};
+
+use crate::jolt::instruction::{
+    prefixes::{PrefixEval, Prefixes},
+    suffixes::{SuffixEval, Suffixes},
+    JoltInstruction,
+};
+use crate::{
+    jolt::{
+        instruction::SubtableIndices,
+        subtable::{eq::EqSubtable, ltu::LtuSubtable, LassoSubtable},
+    },
+    utils::instruction_utils::chunk_and_concatenate_operands,
+};
+
+#[derive(Copy, Clone, Default, Debug, Serialize, Deserialize, PartialEq)]
+pub struct MaxInstruction<const WORD_SIZE: usize>(pub u64, pub u64);
+
+impl<const WORD_SIZE: usize> JoltInstruction for MaxInstruction<WORD_SIZE> {
+    fn operands(&self) -> (u64, u64) {
+        (self.0, self.1)
+    }
+
+    fn combine_lookups<F: JoltField>(&self, vals: &[F], C: usize, M: usize) -> F {
+        let vals_by_subtable = self.slice_values(vals, C, M);
+
+        let left_msb = vals_by_subtable[0];
+        let right_msb = vals_by_subtable[1];
+        let ltu = vals_by_subtable[2];
+        let eq = vals_by_subtable[3];
+        let lt_abs = vals_by_subtable[4];
+        let eq_abs = vals_by_subtable[5];
+
+        // Accumulator for LTU(x_{<s}, y_{<s})
+        let mut ltu_sum = lt_abs[0];
+        // Accumulator for EQ(x_{<s}, y_{<s})
+        let mut eq_prod = eq_abs[0];
+
+        for i in 0..C - 2 {
+            ltu_sum += ltu[i] * eq_prod;
+            eq_prod *= eq[i];
+        }
+        // Do not need to update `eq_prod` for the last iteration
+        ltu_sum += ltu[C - 2] * eq_prod;
+
+        // x_s * (1 - y_s) + EQ(x_s, y_s) * LTU(x_{<s}, y_{<s})
+        let lt = left_msb[0] * (F::one() - right_msb[0])
+            + (left_msb[0] * right_msb[0] + (F::one() - left_msb[0]) * (F::one() - right_msb[0]))
+                * ltu_sum;
+
+        // # Note: we do not need to lookup these values as (limb decompositions of) X and Y will be committed individually anyway
+        let x = F::from_u64(self.0);
+        let y = F::from_u64(self.1);
+        (F::one() - lt) * x + lt * y
+    }
+
+    fn g_poly_degree(&self, C: usize) -> usize {
+        C + 1
+    }
+
+    fn subtables<F: JoltField>(
+        &self,
+        C: usize,
+        _: usize,
+    ) -> Vec<(Box<dyn LassoSubtable<F>>, SubtableIndices)> {
+        vec![
+            (Box::new(LeftMSBSubtable::new()), SubtableIndices::from(0)),
+            (Box::new(RightMSBSubtable::new()), SubtableIndices::from(0)),
+            (Box::new(LtuSubtable::new()), SubtableIndices::from(1..C)),
+            (Box::new(EqSubtable::new()), SubtableIndices::from(1..C - 1)),
+            (Box::new(LtAbsSubtable::new()), SubtableIndices::from(0)),
+            (Box::new(EqAbsSubtable::new()), SubtableIndices::from(0)),
+        ]
+    }
+
+    fn to_indices(&self, C: usize, log_M: usize) -> Vec<usize> {
+        chunk_and_concatenate_operands(self.0, self.1, C, log_M)
+    }
+
+    fn materialize_entry(&self, _: u64) -> u64 {
+        unimplemented!("for shout we will use a simpler min instruction implementation.")
+    }
+
+    fn lookup_entry(&self) -> u64 {
+        if WORD_SIZE == 32 {
+            let x = self.0 as i32;
+            let y = self.1 as i32;
+            x.max(y) as u32 as u64
+        } else if WORD_SIZE == 64 {
+            let x = self.0 as i64;
+            let y = self.1 as i64;
+            x.max(y) as u64
+        } else {
+            panic!("MaxInstruction is only implemented for 32-bit or 64-bit word sizes")
+        }
+    }
+
+    fn random(&self, rng: &mut StdRng) -> Self {
+        match WORD_SIZE {
+            #[cfg(test)]
+            8 => Self(rng.next_u64() % (1 << 8), rng.next_u64() % (1 << 8)),
+            32 => Self(rng.next_u32() as u64, rng.next_u32() as u64),
+            64 => Self(rng.next_u64(), rng.next_u64()),
+            _ => panic!("{WORD_SIZE}-bit word size is unsupported"),
+        }
+    }
+
+    fn evaluate_mle<F: JoltField>(&self, _: &[F]) -> F {
+        unimplemented!("for shout we will use a simpler min instruction implementation.")
+    }
+}
+
+impl<const WORD_SIZE: usize> PrefixSuffixDecomposition<WORD_SIZE> for MaxInstruction<WORD_SIZE> {
+    fn suffixes(&self) -> Vec<Suffixes> {
+        vec![Suffixes::One, Suffixes::LessThan]
+    }
+
+    fn combine<F: JoltField>(&self, prefixes: &[PrefixEval<F>], suffixes: &[SuffixEval<F>]) -> F {
+        debug_assert_eq!(self.suffixes().len(), suffixes.len());
+        let [one, less_than] = suffixes.try_into().unwrap();
+        prefixes[Prefixes::LessThan] * one + prefixes[Prefixes::Eq] * less_than
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use ark_bn254::Fr;
+    use ark_std::test_rng;
+    use rand_chacha::rand_core::RngCore;
+
+    use crate::{
+        jolt::instruction::{
+            test::{
+                instruction_mle_full_hypercube_test, instruction_mle_random_test,
+                materialize_entry_test, prefix_suffix_test,
+            },
+            JoltInstruction,
+        },
+        jolt_instruction_test,
+    };
+
+    use super::MaxInstruction;
+
+    #[test]
+    fn max_instruction_32_e2e() {
+        let mut rng = test_rng();
+        const C: usize = 4;
+        const M: usize = 1 << 16;
+        const WORD_SIZE: usize = 32;
+
+        // Random
+        for _ in 0..256 {
+            let (x, y) = (rng.next_u32() as u64, rng.next_u32() as u64);
+            let instruction = MaxInstruction::<WORD_SIZE>(x, y);
+            jolt_instruction_test!(instruction);
+        }
+
+        // // x == y
+        // for _ in 0..256 {
+        //     let x = rng.next_u32() as u64;
+        //     jolt_instruction_test!(MaxInstruction::<WORD_SIZE>(x, x));
+        // }
+
+        // // Edge cases
+        // let u32_max: u64 = u32::MAX as u64;
+        // let instructions = vec![
+        //     MaxInstruction::<WORD_SIZE>(100, 0),
+        //     MaxInstruction::<WORD_SIZE>(0, 100),
+        //     MaxInstruction::<WORD_SIZE>(1, 0),
+        //     MaxInstruction::<WORD_SIZE>(0, u32_max),
+        //     MaxInstruction::<WORD_SIZE>(u32_max, 0),
+        //     MaxInstruction::<WORD_SIZE>(u32_max, u32_max),
+        //     MaxInstruction::<WORD_SIZE>(u32_max, 1 << 8),
+        //     MaxInstruction::<WORD_SIZE>(1 << 8, u32_max),
+        // ];
+        // for instruction in instructions {
+        //     jolt_instruction_test!(instruction);
+        // }
+    }
+
+    #[test]
+    fn max_instruction_64_e2e() {
+        let mut rng = test_rng();
+        const C: usize = 8;
+        const M: usize = 1 << 16;
+        const WORD_SIZE: usize = 64;
+
+        // Random
+        for _ in 0..256 {
+            let (x, y) = (rng.next_u64(), rng.next_u64());
+            let instruction = MaxInstruction::<WORD_SIZE>(x, y);
+            jolt_instruction_test!(instruction);
+        }
+
+        // x == y
+        for _ in 0..256 {
+            let x = rng.next_u64();
+            jolt_instruction_test!(MaxInstruction::<WORD_SIZE>(x, x));
+        }
+
+        // Edge cases
+        let u64_max: u64 = u64::MAX;
+        let instructions = vec![
+            MaxInstruction::<WORD_SIZE>(100, 0),
+            MaxInstruction::<WORD_SIZE>(0, 100),
+            MaxInstruction::<WORD_SIZE>(1, 0),
+            MaxInstruction::<WORD_SIZE>(0, u64_max),
+            MaxInstruction::<WORD_SIZE>(u64_max, 0),
+            MaxInstruction::<WORD_SIZE>(u64_max, u64_max),
+            MaxInstruction::<WORD_SIZE>(u64_max, 1 << 32),
+            MaxInstruction::<WORD_SIZE>(1 << 32, u64_max),
+            MaxInstruction::<WORD_SIZE>(1 << 63, 1 << (63 - 1)),
+            MaxInstruction::<WORD_SIZE>(1 << (63 - 1), 1 << 63),
+        ];
+        for instruction in instructions {
+            jolt_instruction_test!(instruction);
+        }
+    }
+}
