+++ jolt-core/src/jolt_onnx/instruction/mod.rs
@@ -1,5 +1,5 @@
 //! This module provides the custom jolt instructions for the ONNX runtime.
 
-pub mod min;
+pub mod max;
 pub mod relu;
 pub mod sigmoid;
