+++ jolt-core/src/jolt_onnx/subtable/mod.rs
@@ -1,8 +1,6 @@
 //! This module provides the subtables used by the ONNX runtime.
 
-pub mod higher;
 pub mod is_max;
 pub mod is_pos;
 pub mod is_zero;
-pub mod lower;
 pub mod sigmoid;
